BenDoerr.IO.Common

Description

Common things that I usualy do with IO, error handling and reading files and standared input. Things that I tend to write again and again.

Dealing with IO errors

ioHandled :: IO () -> IO ()

Catch IOErrors using the commonErrorHandler.

Use on main functions for consistent handling of IOErrors at the top level.

 main = ioHandled $ do
                    c <- getContents
                    putStrLn c

commonErrorHandler :: IOError -> IO ()

Handle top level IOErrors consistently.

Currently only handles DoesNotExistErrors.

Will continue to add as I run into them.

Getting Text

readFromStdinOrArgsFiles :: IO ByteString

Wrapper around readFromStdinOrFiles that passes arguments from getArgs.

readFromStdinOrFiles :: [FilePath] -> IO ByteString

If an empty list is given then read from Stdin. Otherwise try reading from files specified by each of the argument list, then join them togeather.

This should read multiple files using near constant memory, there will be a little bump for each addtional file due to an additional file handle.