BenDoerr.WordFrequency.Count

Description

Module BenDoerr.WordFrequency.Count

Description

Types

type WordCounts = Map Word Integer

Simple type alais for our map of words and their counts.

data WordCountDetails

The data type that will hold details about the counted words, including the maximum count and maximum length of a word. These details will be useful when displaying.

The details have been made to evaulate strictly to avoid a buildup of thunks.

Constructors

WordCountDetails

Fields

wordCounts :: !WordCounts

The current state of all counted words. Strict.

maxCount :: !Integer

The current maxiumum count. Strict.

minCount :: !Integer

The current miniumum count. Strict.

maxLength :: !Integer

The current maximum word length. Strict.

Primary Interface

countWords :: [Word] -> WordCountDetails

Builds up WordCountDetails for every word. I think there will be a lot for the GC to do here since we build a new WordCountDetails on every word.

Helpers

addWordOccurrence :: WordCountDetails -> Word -> WordCountDetails

The beef. Gets a new WordCountDetails.

Using strict version of insertWith insertWith' to avoid thunks (slightly better memory footprint).

wordCountsToList :: WordCounts -> [(Word, Integer)]

Transform the WordCounts into a List of (Word, Integer)

The list will be orded descending by the count.