BenDoerr.WordFrequency.Printer

Description

Prints a histogram of the word counts.

TODO: Refactor stats into their own data type.

Print a Histogram

printHistogram

Arguments

:: Integer
Total width of the histogram.

-> WordCountDetails
WordCountDetails to print.

-> IO ()

Main function to print a histogram of the WordCountDetails. This will also print a header out for the histogram with the min and max counts as scale.

Helpers

printWord

Arguments

:: Integer
Total width of the histogram.

-> Integer
Highest count. Max Count.

-> Integer
Lowest count. Min Count.

-> Integer
Length of max word.

-> (Word, Integer)
The word and it's count.

-> IO ()

Print the label and bar for a single word.

The format is: label : #########

findWidth

Arguments

:: Integral b

=> Integer
Margin used when calculating size that can be used.

-> Integer
Total width of the histogram.

-> Integer
Max Count.

-> Integer
Min Count.

-> Integer
The current count.

-> b

Find the width to use for the bar given the max size of the histogram, the max size of the label and any margin between the label and bar.

printHeader

Arguments

:: Integer
Total width of the histogram.

-> Integer
Max Count.

-> Integer
Min Count.

-> Integer
Length of max word.

-> IO ()

Print a nice header with some bold borders and show the minimum count and maximum count for scale and perspective.