BenDoerr.WordFrequency.Reader

Description

Module BenDoerr.WordFrequency.Reader

Description

Primary Interface

getSpecimen :: IO [Word]

The primary function that takes input from a list of files or if no files are specified then from stdin. Breaks the text in those files into their word. Words are normalized in lower case and have surrounding punctuation removed.

type Word = Text

Type alias of Text for a type that represents a word from the specimen.

Building the Word List

cleanWords :: ByteString -> [Word]

Read the ByteString as UTF-8, separate it into words, lowercase the words, drop punctuation around the word, and return only non null words.

normalizeWord :: Word -> Word

Converts the word to lowercase and removes surrounding punctuation.

Punctuation is considered *,.-!?<>/";:][}{)(@#$%^&`~-_=+

Helpers

toString :: Word -> String

Unpack a Word to a String.

wordLength :: Word -> Integer

The length of a particular word.

printSpecimen :: [Word] -> IO ()

Print a word list.

printSpecimen' :: IO ()

Print a word list that is the result of getSpecimen